requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-filter==22.1
django-import-export==3.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
future==0.18.2
idna==3.4
iniconfig==1.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
packaging==22.0
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.3
tablib==3.3.0
toml==0.10.2
urllib3==1.26.13
xlrd==2.0.1
xlwt==1.3.0
